fix(api): 修复 roleID 类型不匹配导致权限加载失败
问题：admin 用户只能看到 Dashboard 和 Logs 菜单项，
其他菜单项因权限数组为空而隐藏。

原因：
- JWTClaims.RoleID 是 int32 类型
- middleware 将 int32 存入 context
- pkg/auth/context.go GetUserRoleID 期望 int64 类型断言
- 类型断言失败，返回错误，permissions 数组为空

修复：
- GetUserRoleID 返回类型改为 int32
- auth handler Me 函数移除多余的类型转换

+++ api/internal/handler/auth.go
@@ -397,7 +397,7 @@ func (h *AuthHandler) Me(w http.ResponseWriter, r *http.Request) {
 				permissions = append(permissions, code)
 			}
 		} else if err == nil {
-			permissions = h.permService.GetPermissionsByRoleID(int32(roleID))
+			permissions = h.permService.GetPermissionsByRoleID(roleID)
 		}
 	}
